Fix purge and App functionality
Fix purge because it wasn't taking into account the node and that the pop() over the Stack structure could modify the number of iterations to remove the corresponding domNodes position correctly. Fix the App to include localSearch and purge for each local solution, and not the best one the builder found.